pyproject.toml
[tool.poetry]
name = "fds.sdk.utils"
version = "0.9.0"
description = "Utilities for interacting with FactSet APIs."
authors=["FactSet Research Systems"]
license="Apache-2.0"
readme="README.md"
homepage="https://developer.factset.com"
keywords=[
  "FactSet",
  "API",
  "SDK"
]
packages=[
  { include = "fds", from = "src" }
]
exclude=["tests"]

[tool.poetry.dependencies]
python = "^3.6.2"
python-jose = "^3.3.0"
requests-oauthlib = "^1.3.0"
requests = "^2.25.1"
oauthlib = "^3.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.6b0", allow-prereleases = true}
pytest-cov = "^2.12.1"
pytest-mock = "^3.6.1"
tox = "^3.24.1"
tox-gh-actions = "^2.6.0"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
